Rename RootStoreContext to RootContext
