+++ src/containers/ChannelMessages.js
@@ -7,12 +7,12 @@ import classNames from 'classnames'
 
 import getMousePosition from '../utils/mouse-position'
 
-import RootStoreContext from '../context/RootStoreContext'
+import RootContext from '../context/RootContext'
 
 import MessageList from '../components/MessageList'
 
 function ChannelMessages ({ channel }) {
-  const { sessionStore, uiStore } = useContext(RootStoreContext)
+  const { sessionStore, uiStore } = useContext(RootContext)
 
   function onMessageUserClick (evt, profile, identity) {
     evt.persist()
